Rename package invalid package name for drivers/restaskdocker
