+++ drivers/restaskdocker/caps.go
@@ -1,4 +1,4 @@
-package restaskpodman
+package restaskdocker
 
 import (
 	"os/exec"
