+++ drivers/restaskdocker/main.go
@@ -1,4 +1,4 @@
-package restaskpodman
+package restaskdocker
 
 // TODO
 // * snooze
